Rename db/companions.js -> db/matrix.js
Now there are the terms "companionship algorithm" and "companion plant
matrix".
